requirements.txt
aiohttp==3.8.5 ; python_version >= "3.9" and python_version < "4.0"
aioredis==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
amqp==5.1.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
billiard==4.1.0 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.28.26 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.31.26 ; python_version >= "3.9" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.9" and python_version < "4.0"
celery==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.4.2 ; python_version >= "3.9" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.101.1 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
iso8601==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
kombu==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
linkify==1.4 ; python_version >= "3.9" and python_version < "4.0"
lxml==4.9.3 ; python_version >= "3.9" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.2 ; python_version >= "3.9" and python_version < "4.0"
openai==0.27.8 ; python_version >= "3.9" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic[email]==1.10.2 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
pypdf==3.15.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
redis==5.0.0 ; python_version >= "3.9" and python_version < "4.0"
regex==2023.8.8 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rich==13.5.2 ; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
tiktoken==0.4.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.9" and python_version < "4.0"
uvicorn[standard]==0.23.2 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.9" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.9" and python_version < "4.0"
